Data.Tuple.Append.Instances.Vector

Description

A module that contains typeclasses instances for the Vector, Vector, Vector, Vector, Bundle, Bundle, and Stream data types.

Orphan instances

TupleAddL a (Vector a) (Vector a)

Instance details

Methods

(<++) :: a -> Vector a -> Vector a

Prim a => TupleAddL a (Vector a) (Vector a)

Instance details

Methods

(<++) :: a -> Vector a -> Vector a

Storable a => TupleAddL a (Vector a) (Vector a)

Instance details

Methods

(<++) :: a -> Vector a -> Vector a

Unbox a => TupleAddL a (Vector a) (Vector a)

Instance details

Methods

(<++) :: a -> Vector a -> Vector a

TupleAddL a (Bundle v a) (Bundle v a)

Instance details

Methods

(<++) :: a -> Bundle v a -> Bundle v a

Monad m => TupleAddL a (Stream m a) (Stream m a)

Instance details

Methods

(<++) :: a -> Stream m a -> Stream m a

Monad m => TupleAddL a (Bundle m v a) (Bundle m v a)

Instance details

Methods

(<++) :: a -> Bundle m v a -> Bundle m v a

TupleAddR (Vector a) a (Vector a)

Instance details

Methods

(++>) :: Vector a -> a -> Vector a

Prim a => TupleAddR (Vector a) a (Vector a)

Instance details

Methods

(++>) :: Vector a -> a -> Vector a

Storable a => TupleAddR (Vector a) a (Vector a)

Instance details

Methods

(++>) :: Vector a -> a -> Vector a

Unbox a => TupleAddR (Vector a) a (Vector a)

Instance details

Methods

(++>) :: Vector a -> a -> Vector a

TupleAppend (Vector a) (Vector a) (Vector a)

Instance details

Methods

(+++) :: Vector a -> Vector a -> Vector a

Prim a => TupleAppend (Vector a) (Vector a) (Vector a)

Instance details

Methods

(+++) :: Vector a -> Vector a -> Vector a

Storable a => TupleAppend (Vector a) (Vector a) (Vector a)

Instance details

Methods

(+++) :: Vector a -> Vector a -> Vector a

Unbox a => TupleAppend (Vector a) (Vector a) (Vector a)

Instance details

Methods

(+++) :: Vector a -> Vector a -> Vector a

TupleAddR (Bundle v a) a (Bundle v a)

Instance details

Methods

(++>) :: Bundle v a -> a -> Bundle v a

Monad m => TupleAddR (Stream m a) a (Stream m a)

Instance details

Methods

(++>) :: Stream m a -> a -> Stream m a

TupleAppend (Bundle v a) (Bundle v a) (Bundle v a)

Instance details

Methods

(+++) :: Bundle v a -> Bundle v a -> Bundle v a

Monad m => TupleAppend (Stream m a) (Stream m a) (Stream m a)

Instance details

Methods

(+++) :: Stream m a -> Stream m a -> Stream m a

Monad m => TupleAddR (Bundle m v a) a (Bundle m v a)

Instance details

Methods

(++>) :: Bundle m v a -> a -> Bundle m v a

Monad m => TupleAppend (Bundle m v a) (Bundle m v a) (Bundle m v a)

Instance details

Methods

(+++) :: Bundle m v a -> Bundle m v a -> Bundle m v a